feature/prepare-development コンポーネントディレクトリの整理
