+++ components/auth/forgot-password-form.tsx
@@ -12,17 +12,17 @@ import {
 } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import { useRouter } from "next/navigation";
+import Link from "next/link";
 import { useState } from "react";
 
-export function UpdatePasswordForm({
+export function ForgotPasswordForm({
   className,
   ...props
 }: React.ComponentPropsWithoutRef<"div">) {
-  const [password, setPassword] = useState("");
+  const [email, setEmail] = useState("");
   const [error, setError] = useState<string | null>(null);
+  const [success, setSuccess] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
-  const router = useRouter();
 
   const handleForgotPassword = async (e: React.FormEvent) => {
     e.preventDefault();
@@ -31,9 +31,11 @@ export function UpdatePasswordForm({
     setError(null);
 
     try {
-      const { error } = await supabase.auth.updateUser({ password });
+      const { error } = await supabase.auth.resetPasswordForEmail(email, {
+        redirectTo: `${window.location.origin}/auth/update-password`,
+      });
       if (error) throw error;
-      router.push("/dashboard");
+      setSuccess(true);
     } catch (error: unknown) {
       setError(error instanceof Error ? error.message : "An error occurred");
     } finally {
@@ -45,28 +47,33 @@ export function UpdatePasswordForm({
     <div className={cn("flex flex-col gap-6", className)} {...props}>
       <Card>
         <CardHeader>
-          <CardTitle className="text-2xl">パスワードの変更</CardTitle>
+          <CardTitle className="text-2xl">パスワードを忘れた場合</CardTitle>
           <CardDescription>
-            新しいパスワードを入力してください
+            メールアドレスを入力してください。パスワードリセットのリンクを送信します。
           </CardDescription>
         </CardHeader>
         <CardContent>
           <form onSubmit={handleForgotPassword}>
             <div className="flex flex-col gap-6">
               <div className="grid gap-2">
-                <Label htmlFor="password">新しいパスワード</Label>
+                <Label htmlFor="email">メールアドレス</Label>
                 <Input
-                  id="password"
-                  type="password"
-                  placeholder="新しいパスワード"
+                  id="email"
+                  type="email"
+                  placeholder="example@company.com"
                   required
-                  value={password}
-                  onChange={(e) => setPassword(e.target.value)}
+                  value={email}
+                  onChange={(e) => setEmail(e.target.value)}
                 />
               </div>
               {error && <p className="text-sm text-red-500">{error}</p>}
+              {success && (
+                <p className="text-sm text-green-500">
+                  パスワードリセットのリンクを送信しました。メールをご確認ください。
+                </p>
+              )}
               <Button type="submit" className="w-full" disabled={isLoading}>
-                {isLoading ? "保存中..." : "パスワードを保存"}
+                {isLoading ? "送信中..." : "リセットリンクを送信"}
               </Button>
             </div>
           </form>
@@ -74,4 +81,4 @@ export function UpdatePasswordForm({
       </Card>
     </div>
   );
-}
+}
