+++ components/auth/logout-button.tsx
@@ -13,5 +13,5 @@ export function LogoutButton() {
     router.push("/auth/login");
   };
 
-  return <Button onClick={logout}>Logout</Button>;
-}
+  return <Button onClick={logout}>ログアウト</Button>;
+}
